chore(dev.Dockerfile): update Dockerfile to set NO_FRONTEND environment variable and modify start script to conditionally build frontend based on the presence of the variable

+++ docker/dev.start.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
-cd src/frontend \
-    && rm -rf node_modules \
-    && npm install \
-    && npm run dev:docker &
+# Check if frontend should be built
+if [ "$1" != "no-frontend" ]; then
+    cd src/frontend \
+        && rm -rf node_modules \
+        && npm install \
+        && npm run dev:docker &
+fi
+
 make backend
